+++ adminer/include/html.inc.php
@@ -426,7 +426,7 @@ function search_tables(): void {
 	foreach (table_status('', true) as $table => $table_status) {
 		$name = adminer()->tableName($table_status);
 		if (isset($table_status["Engine"]) && $name != "" && (!$_POST["tables"] || in_array($table, $_POST["tables"]))) {
-			$result = connection()->query("SELECT" . limit("1 FROM " . table($table), " WHERE " . implode(" AND ", adminer()->selectSearchProcess(fields($table), array())), 1));
+			$result = connection()->query("SELECT" . limit("1 FROM " . table($table), " WHERE " . implode(" AND ", adminer()->selectSearchProcess(fields($table), array(), $table_status)), 1));
 			if (!$result || $result->fetch_row()) {
 				$print = "<a href='" . h(ME . "select=" . url_escape($table)
 					. "&where[0][op]=" . url_escape($_GET["where"][0]["op"])